+++ src/common/types.ts
@@ -1,1 +1,6 @@
 export type GenericDataType = { id: string | number } & Record<string, unknown>;
+
+export interface Transform<TOutput> {
+  output: (val: string) => TOutput | undefined;
+  input: (val: TOutput) => string;
+}